+++ __tests__/src/test-utils/setup-test-db.ts
@@ -83,7 +83,7 @@ const outboxSetup = async (
         id uuid PRIMARY KEY,
         aggregate_type VARCHAR(255) NOT NULL,
         aggregate_id VARCHAR(255) NOT NULL,
-        event_type VARCHAR(255) NOT NULL,
+        message_type VARCHAR(255) NOT NULL,
         payload JSONB NOT NULL,
         created_at TIMESTAMPTZ NOT NULL DEFAULT NOW()
       );
@@ -134,7 +134,7 @@ const inboxSetup = async (
         id uuid PRIMARY KEY,
         aggregate_type VARCHAR(255) NOT NULL,
         aggregate_id VARCHAR(255) NOT NULL,
-        event_type VARCHAR(255) NOT NULL,
+        message_type VARCHAR(255) NOT NULL,
         payload JSONB NOT NULL,
         created_at TIMESTAMPTZ NOT NULL,
         processed_at TIMESTAMPTZ,