+++ examples/rabbitmq/consumer/setup/init-db.ts
@@ -84,7 +84,7 @@ const inboxSetup = async (config: Config): Promise<void> => {
         id uuid PRIMARY KEY,
         aggregate_type VARCHAR(255) NOT NULL,
         aggregate_id VARCHAR(255) NOT NULL, 
-        event_type VARCHAR(255) NOT NULL,
+        message_type VARCHAR(255) NOT NULL,
         payload JSONB NOT NULL,
         created_at TIMESTAMPTZ NOT NULL,
         processed_at TIMESTAMPTZ,