+++ examples/rabbitmq/producer/src/add-movies.ts
@@ -8,11 +8,11 @@ import { Config } from './config';
 import { logger } from './logger';
 
 export const MovieAggregateType = 'movie';
-export const MovieCreatedEventType = 'movie_created';
+export const MovieCreatedMessageType = 'movie_created';
 
 /**
  * A business logic function to add business data (movies) to the database and
- * register an outbox message event "movie_created" for every inserted movie.
+ * register an outbox message "movie_created" for every inserted movie.
  * It uses a small timeout to insert a movie and an outbox message every
  * second.
  * @param config The configuration object with details on how to connect to the database with the login role.
@@ -32,10 +32,10 @@ export const addMovies = async (
     logger.error(err, 'PostgreSQL pool error');
   });
 
-  // Create the outbox storage function for the movie created event
+  // Create the outbox storage function for the movie created event message
   const storeOutboxMessage = initializeOutboxMessageStorage(
     MovieAggregateType,
-    MovieCreatedEventType,
+    MovieCreatedMessageType,
     outboxConfig,
   );
 
@@ -61,7 +61,7 @@ const insertMovie = async (dbClient: PoolClient) => {
   if (movieInsertedIdResult.rowCount === 0) {
     throw new Error('Could not insert the movie.');
   }
-  // Selecting a few properties to send as part of the event
+  // Selecting a few properties to send as part of the message
   const createdMovie = movieInsertedIdResult.rows[0];
   logger.trace(createdMovie, 'Stored a movie');
   return createdMovie;