+++ examples/rabbitmq/producer/src/rabbitmq-publisher.ts
@@ -27,7 +27,7 @@ export const initializeRabbitMqPublisher = async (
     async (message: OutboxMessage): Promise<void> => {
       // Publish a message
       const publication = await broker.publish(
-        message.eventType, // By convention we use the event type also as publish topic
+        message.messageType, // By convention we use the message type also as publish topic
         message, // Send the full outbox message so a receiver can use the inbox pattern to check for duplicate messages
       );
       return new Promise((resolve, reject) => {