+++ lib/src/inbox.ts
@@ -128,15 +128,15 @@ const insertInbox = async (
   dbClient: PoolClient,
   { settings }: Pick<InboxServiceConfig, 'settings'>,
 ) => {
-  const { id, aggregateType, aggregateId, eventType, payload, createdAt } =
+  const { id, aggregateType, aggregateId, messageType, payload, createdAt } =
     message;
   const inboxResult = await dbClient.query(
     /* sql*/ `
     INSERT INTO ${settings.dbSchema}.${settings.dbTable}
-      (id, aggregate_type, aggregate_id, event_type, payload, created_at)
+      (id, aggregate_type, aggregate_id, message_type, payload, created_at)
       VALUES ($1, $2, $3, $4, $5, $6)
       ON CONFLICT (id) DO NOTHING`,
-    [id, aggregateType, aggregateId, eventType, payload, createdAt],
+    [id, aggregateType, aggregateId, messageType, payload, createdAt],
   );
   if (inboxResult.rowCount < 1) {
     logger().warn(message, `The message with id ${id} already existed`);